Incremental commit on dem Xarray accessor
