+++ setup.cfg
@@ -46,6 +46,7 @@ xdem =
 include =
     xdem
     xdem.coreg
+    xdem.dem
 
 [options.extras_require]
 opt =
