+++ tests/test_dem/test_xr_accessor.py
@@ -1,0 +1,6 @@
+"""Test module for 'dem' Xarray accessor."""
+
+class TestDEMAccessor:
+
+    def test_open_dem(self):
+        pass
