+++ xdem/dem/__init__.py
@@ -1,0 +1,1 @@
+from xdem.dem.dem import *  # noqa
